requirements.txt
fair-esm
torch
flask
